refactor some files about ScheduleSection
